# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SCADA Logistics API",
    "description": "Мониторинг логистической инфраструктуры предприятия.\n\n**Авторизация:** `POST /auth/login` → `Authorization: Bearer <access_token>`; access живёт 15 мин, обновление — `POST /auth/refresh`. WebSocket: `/ws/live?token=<access_token>`. Демо-пользователи: `dispatcher`, `security`, `admin`, пароль `demo`.\n\nКоординаты — метры локального плана (см. `GET /objects`). Время — ISO 8601 с таймзоной.\n\nWebSocket `/ws/live`: формат сообщений — `contracts/ws.schema.json`.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/sensor-types": {
      "get": {
        "tags": [
          "registry"
        ],
        "summary": "List Sensor Types",
        "operationId": "list_sensor_types_sensor_types_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SensorTypeInfo"
                  },
                  "type": "array",
                  "title": "Response List Sensor Types Sensor Types Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/sensors": {
      "get": {
        "tags": [
          "registry"
        ],
        "summary": "List Sensors",
        "operationId": "list_sensors_sensors_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SensorType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "name": "building_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Building Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Sensor"
                  },
                  "title": "Response List Sensors Sensors Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "registry"
        ],
        "summary": "Create Sensor",
        "description": "Без `thresholds` датчик получает типовые пороги своего типа. Начинает принимать данные сразу (коннекторы\nперечитывают реестр), симулятор подхватывает климат и движение в течение минуты.",
        "operationId": "create_sensor_sensors_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SensorCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sensor"
                }
              }
            }
          },
          "409": {
            "description": "Sensor id already exists"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sensors/{sensor_id}": {
      "get": {
        "tags": [
          "registry"
        ],
        "summary": "Get Sensor",
        "operationId": "get_sensor_sensors__sensor_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sensor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sensor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sensor"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "registry"
        ],
        "summary": "Update Sensor",
        "operationId": "update_sensor_sensors__sensor_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sensor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sensor Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SensorUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sensor"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sensors/{sensor_id}/thresholds": {
      "put": {
        "tags": [
          "registry"
        ],
        "summary": "Set Thresholds",
        "description": "Новые пороги становятся следующей версией; старые закрываются (`valid_to`). Поле `version` во входе игнорируется.\nWorker перечитывает пороги каждые 30 с.",
        "operationId": "set_thresholds_sensors__sensor_id__thresholds_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sensor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sensor Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Threshold"
                },
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sensor"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sensors/bulk": {
      "post": {
        "tags": [
          "registry"
        ],
        "summary": "Bulk Create Sensors",
        "description": "CSV (колонки как у SensorCreate; `x`,`y` — координаты; `thresholds` — JSON-строка) или JSON-массив.\n\nВсё или ничего: при любой ошибке не создаётся ни один датчик, ошибки возвращаются построчно.",
        "operationId": "bulk_create_sensors_sensors_bulk_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dry_run",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Только проверить, ничего не создавать",
              "default": true,
              "title": "Dry Run"
            },
            "description": "Только проверить, ничего не создавать"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_bulk_create_sensors_sensors_bulk_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sensors/{sensor_id}/history": {
      "get": {
        "tags": [
          "registry"
        ],
        "summary": "Sensor History",
        "description": "Шаг выбирается по длине периода: до 1 ч — сырые точки, до суток — минутные агрегаты, дальше — часовые.",
        "operationId": "sensor_history_sensors__sensor_id__history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sensor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sensor Id"
            }
          },
          {
            "name": "metric",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Числовая метрика типа датчика, напр. temperature_c",
              "title": "Metric"
            },
            "description": "Числовая метрика типа датчика, напр. temperature_c"
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "По умолчанию — час назад",
              "title": "From"
            },
            "description": "По умолчанию — час назад"
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "По умолчанию — сейчас",
              "title": "To"
            },
            "description": "По умолчанию — сейчас"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SensorHistory"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/vehicles": {
      "get": {
        "tags": [
          "registry"
        ],
        "summary": "List Vehicles",
        "operationId": "list_vehicles_vehicles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Vehicle"
                  },
                  "type": "array",
                  "title": "Response List Vehicles Vehicles Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/vehicles/{vehicle_id}/route": {
      "get": {
        "tags": [
          "registry"
        ],
        "summary": "Vehicle Route",
        "description": "Трек машины по ГЛОНАСС; длинные периоды прореживаются до ~2000 точек (последняя точка в интервале).",
        "operationId": "vehicle_route_vehicles__vehicle_id__route_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "vehicle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Vehicle Id"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "По умолчанию — час назад",
              "title": "From"
            },
            "description": "По умолчанию — час назад"
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "По умолчанию — сейчас",
              "title": "To"
            },
            "description": "По умолчанию — сейчас"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleRoute"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/whitelist": {
      "get": {
        "tags": [
          "registry"
        ],
        "summary": "List Whitelist",
        "description": "Пропуска и номера. ФИО видны только с правом `people:view_pii` (охрана, админ), остальным — маскированные.",
        "operationId": "list_whitelist_whitelist_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(card|plate)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Поиск по номеру пропуска / гос. номеру",
              "title": "Q"
            },
            "description": "Поиск по номеру пропуска / гос. номеру"
          },
          {
            "name": "active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Active"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhitelistPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_bulk_create_sensors_sensors_bulk_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_bulk_create_sensors_sensors_bulk_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SensorType": {
        "type": "string",
        "enum": [
          "anpr_camera",
          "access_control",
          "gnss",
          "motion",
          "climate"
        ],
        "title": "SensorType"
      },
      "SensorTypeInfo": {
        "properties": {
          "id": {
            "$ref": "#/components/schemas/SensorType"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "Климат (температура и влажность)"
            ]
          },
          "is_mobile": {
            "type": "boolean",
            "title": "Is Mobile",
            "description": "Подвижный датчик: позиция приходит в событиях, а не задана на плане"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/MetricSpec"
            },
            "type": "array",
            "title": "Metrics"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "is_mobile",
          "metrics"
        ],
        "title": "SensorTypeInfo"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}